# implements ScorePlugin