--- a/Makefile
+++ b/Makefile
@@ -6,7 +6,7 @@ golangci_lint := github.com/golangci/golangci-lint/cmd/golangci-lint@v1.53.2
 	@(cd $(@D); tinygo build -o main.wasm -gc=custom -tags=custommalloc -scheduler=none --no-debug -target=wasi main.go)
 
 .PHONY: build-tinygo
-build-tinygo: examples/filter-simple/main.wasm examples/noop/main.wasm
+build-tinygo: examples/filter-simple/main.wasm examples/score-simple/main.wasm guest/testdata/noop/main.wasm
 
 %/main-debug.wasm: %/main.go
 	@(cd $(@D); tinygo build -o main-debug.wasm -gc=custom -tags=custommalloc -scheduler=none -target=wasi main.go)