#### What type of PR is this?

/kind feature

#### What this PR does / why we need it:

This implements ScorePlugin. Notably, this reduces the size of the score returnable by wasm to int32 from int64 for reasons elaborated in the RATIONALE.md.

#### Which issue(s) this PR fixes:
<!--
*Automatically closes linked issue when PR is merged.
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
_If PR is about `failing-tests or flakes`, please post the related issues/tests in a comment and do not use `Fixes`_*
-->
Fixes #

#### Special notes for your reviewer:

Thanks a lot to @anuraaga and @ncruces for helping with the wasm aspects of the rationale.

Thanks a lot to @kerthcet and @sanposhiho for helping with the value space aspects of the score plugin.

#### Does this PR introduce a user-facing change?

Only notable behavior is reducing the value range of score produced by wasm plugins to max int32.

#### What are the benchmark results of this change?

```
goos: darwin
goarch: arm64
pkg: sigs.k8s.io/kube-scheduler-wasm-extension/internal/e2e
                                            │   old.txt   │               new.txt               │
                                            │   sec/op    │   sec/op     vs base                │
PluginFilter/noop/params:_small-12            179.5n ± 0%   181.0n ± 0%  +0.84% (p=0.002 n=6)
PluginFilter/noop/params:_real-12             178.8n ± 1%   182.3n ± 1%  +1.96% (p=0.004 n=6)
PluginFilter/filter-simple/params:_small-12   6.750µ ± 2%
PluginFilter/filter-simple/params:_real-12    50.62µ ± 2%
PluginFilter/noop-wat/params:_small-12                      120.1n ± 0%
PluginFilter/noop-wat/params:_real-12                       119.8n ± 0%
PluginFilter/wasm/params:_small-12                          6.671µ ± 1%
PluginFilter/wasm/params:_real-12                           50.26µ ± 2%
PluginScore/noop-wat/params:_small-12                       120.5n ± 2%
PluginScore/noop-wat/params:_real-12                        122.0n ± 1%
PluginScore/noop/params:_small-12                           244.5n ± 4%
PluginScore/noop/params:_real-12                            206.5n ± 1%
PluginScore/wasm/params:_small-12                           5.082µ ± 5%
PluginScore/wasm/params:_real-12                            49.80µ ± 1%
geomean                                       1.820µ        746.4n       +1.39%               ¹
¹ benchmark set differs from baseline; geomeans may not be comparable
```