@codefromthecrypt 2023-06-09
> In theory, yes, we have no idea how people design their plugins. Although after normalizing, they will be in the range of [0, 100]

ok I will put the note in that normalizing means [0, 100] so basically regardless of number above 100 it normalizes to 100.