@codefromthecrypt 2023-06-09
So, next steps for me are:
* make sure the score and status results are signed, even if narrowed to 32-bits. This increases parity even, even if negative values aren't semantically defined for either.
* consolidate confusingly large types mapped to also too large types, but more convenient ones, into a section. It seems a trend to accidentally (int64) or by default (int) use really large types for values never intended to be large).    
  * This allows us to refer to that section in places handling type differences.
  * Add to this section that especially in score it is unnecessary as the usable range is [0, 100]
  * Add to this section that the user can only define 2 million custom positive status codes now ;)
* restructure the notes currently specifically about score into a general note about when returning two values.
* fill some tests.

Will push a commit to include this!