@codefromthecrypt 2023-06-10
ok spent another hour plus on docs. I think that's way over budget for me and unparalleled in practice.

Next is test.