@codefromthecrypt 2023-06-10
ok addressed all the TODO I'm aware of, PTAL!